Make base for tests
